backend/requirements.txt
# FastAPI Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Geospatial
rasterio>=1.3.9
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
osmnx>=1.9.0
scipy>=1.12.0
numpy>=1.26.0

# Image processing
Pillow>=10.0.0
matplotlib>=3.8.0

# Caching & Performance
cachetools>=5.3.0

# HTTP Client (for AQI API)
httpx>=0.26.0

# Database (for community suggestions)
pymongo>=4.6.0
dnspython

# CORS
aiofiles>=23.2.0

# Type hints
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Telegram Bot
python-telegram-bot>=21.0
